backend/requirements.txt
firebase_admin
fastapi
uvicorn
asyncio
pymongo
pydantic
python-dotenv